Fix usages of Jimfs in tests on Windows.
Jimfs defaults to using a configuration that matches the OS it is
running on. On Windows, that means that there is no such thing as a
top-level `/tmp` path. Instead, we now use
`getPath("tmp").toAbsolutePath()`.

We add some smoke tests in the AppVeyor build to prevent
regressions in the future.