Remove the polymorphic Array/TypedArray split in primitive arrays.
The intrinsics for Array to TypedArray conversions were creating
instances of our `ArrayClass`es with typed arrays instead of Arrays
in the `u` field. This was intended as an optimization, but we
never measured that. It may even trigger performance cliffs due to
the fact that it makes the `u` file polymorphic instead of
monomorphic.