Fix #4325: Fix GCC pretty printing
Our usage of `SourceFile` was wrong: It has provisions to retrieve the
original source code and some of the pretty printing functionality
attempts to use this (e.g. to write a number as it was written in the
original source).

We switch to `SimpleSourceFile` which contains a minimum
implementation of `StaticSourceFile` (the interface actually required
by `Node`).