Fix #4328: Fix Class.isAssignableFrom for hijacked classes and JS types.
Previously, we were creating a "fake instance" of the argument
class, and passing it to `isInstance`. This caused a number of
issues for hijacked classes and JS types, as detailed in the issue.

We fix all those issues by not relying on fake instances anymore.
Instead, we directly compare the class data, notably using the
`ancestors` field.

The new comparison matches the one used to test instances of
arrays, which is natural since they now both correspond to the JVM
subtyping test. We therefore rearrange the `isInstance` test of
array classes to build on the same infrastructure as
`isAssignableFrom`. This allows to simplify the class data in the
process.

The changes do not impact the results for `Class`es representing
array classes. Those were already in sync with the instance tests
for arrays, and with the JVM behavior.