Consolidate and expand tests for run-time type tests.
Previously, tests for `isInstanceOf`, `asInstanceOf`,
`jl.Class.isInstance` and `jl.Class.cast` were scattered over
several files, with varying coverage of source and target types.

In this commit, we consolidate the tests for all of those type
tests in a pair of test files `RuntimeTypeTestsTest.scala` and
`RuntimeTypeTestsJSTest.scala`. We use a shared infrastructure to
consistently test all four run-time type tests.

In addition, we expand the coverage of tests, especially for array
types.