Encode arrays of primitive types as typed arrays when possible.
We use the following typed arrays as storage:

* `Array[Char]`: `Uint16Array`
* `Array[Byte]`: `Int8Array`
* `Array[Short]`: `Int16Array`
* `Array[Int]`: `Int32Array`
* `Array[Float]`: `Float32Array`
* `Array[Double]`: `Float64Array`

In addition, when `Long`s are encoded as `bigint`s, we use
`BitInt64Array` for `Array[Long]`.

Other array classes still use a standard JavaScript `Array` as
their underlying storage. In particular, `Array[Boolean]` does,
although we could consider using a `Uint8Array` (packed or not) in
the future.

We could also consider using an `Int32Array` with twice the number
of elements for `Array[Long]`, but that requires a deeper
integration with the optimizer to be beneficial.