Introduce a transient ResolvedApply node, inserted by the optimizer.
When the optimizer resolves an `Apply` or an `ApplyStatically` to a
single target, but does not inline it, it used to throw out the
knowledge and re-inserted the same call. Now, it instead inserts
a `ResolvedApply`, which retains the precise target.

In this commit, the emitter is adapted to pass through such nodes,
ignoring this additional knowledge. The ultimate output is
therefore not changed at this point.

The analyzer uses that information, in the sense that some `Apply`
nodes can be looked up as if they were `ApplyStatically`. This
rarely results in any actual change, however. The test suite
contains only 1 method that can be dead-code-eliminated after this
commit but not before.

Other than that, the output is the same.