Add the set of public methods called non-resolved in a LinkedClass.
We refine the Analyzer to track resolved calls differently from
non-resolved calls. The distinction is only meaningful for methods
in the Public namespace. A non-resolved call sets an additional
flag `isReachableNonResolved` on the `MethodInfo`, which records
that it is called in a non-resolved context at least once.

In the `Analysis` and further in the `LinkedClass`, we store the
set of public method names that are only reachable in resolved
contexts.

In a future commit, the emitter will be able to use that
information to compile always-resolved methods in a more efficient
way. This commit, however, does not affect the output in any way.