Emit always-resolved public methods as static-like.
Using the information built in the optimizer and in the refiner's
analysis, we finally change the emitter to optimize calls to public
methods that are always resolved.

If a public method is always called in a `ResolvedApply`, we emit
it as a static-like method, using the same infrastructure as for
default methods and hijacked class methods (which, in a sense, are
always resolved by construction).