Fix #4357: Inc. optimizer: Invalidate callers if `shouldInline` changes.
In `GenIncOptimizer`, we diff the attributes of a `MethodImpl`
before and after a change. When any of those changes, callers need
to be reoptimized. Previously, the comparison was bogus because we
forgot to take the `shouldInline` attribute into account.

We refactor the handling of attributes to make it harder to make
this mistake in the future. We use one case class `Attributes` that
aggregates everything we care about in the superclass
`OptimizerCore.MethodImpl`. The subclass in `GenIncOptimizer` does
not need to know the list of attributes anymore, since it
manipulates an instance of `Attributes` as a whole.