Implement SHA-1 hashing privately in irJS, to make Hashers work.
Previously, although hashing utilities were located in the shared
sources of scalajs-ir, trying to use them on the JS platform would
fail to link, since `java.security.MessageDigest` is not supported.

We introduce platform-specific sources for a SHA-1 hasher, which
uses `java.security.MessageDigest` on the JVM, and an own
implementation on JS. Our implementation is based on an existing
implementation in C, referenced in the comments.

Shared tests ensure that our implementation corresponds to what the
JVM does.

The existing JVM-only tests for `Hashers` are moved to the `shared`
folder.