Fix Hashers.hashAsVersion not to ignore part of the data.
The previous version used `forDigit(b >> 4, 16)`, which would
result in a negative value if the byte `b` was negative. This is an
invalid argument to `forDigit`, which therefore returns the null
character `\u0000`. This would conflate all the high-4-bits of
negative bytes to the same character, creating collisions.

We fix this with `(b >> 4) & 0x0f`.