Upgrade sbt to 1.3.13.
This is the latest (and very likely last) 1.3.x version.

Recent versions of Metals cannot import builds using sbt 1.2.x
anymore. The initial import fails with a LinkageError because it
wants to access a setting that was added in sbt 1.3.0.

This change may make our sbt plugin incompatible with sbt 1.2.x
builds, as our main source of tests for it are our own build.
However, almost all the scripted tests force to use sbt 1.2.8, and
should give some confidence that we aren't breaking anything.

If we do, oh well, sbt 1.2.x is a bit old anyway.