Fix #4372: Generalize js.Function's to allow user-defined ones.
Previously, there was a fixed set of built-in JS function types,
`js.FunctionN` and `js.ThisFunctionN`, that represented JS function
values. The only way to create JS function values was to have a
lambda expression used as a SAM of these types.

In this commit, we generalize these to any SAM-able non-native JS
trait whose single abstract method is named `apply` and whose
super-*class* is `js.Function`. This allows the user to define
custom JS function types for signatures that are not made available
by the built-in types. A typical example is signatures with rest
parameters (i.e., varargs).

To do this, we make the following changes:

* We allow declaring an *abstract* `apply` method in a non-native
  JS trait if that makes that trait a SAM type with that method and
  if its superclass is `js.Function`.
  Even though we cannot in general implement `apply` methods, it is
  only the *concrete* ones that are problematic.
* We apply the JS `Closure` treatment in the back-end to any
  anonymous function class that is the result of desugaring a
  lambda for such a SAM type.
* In that treatment, we make a `function` function (as opposed to
  an arrow function) and use a `this` parameter if the anonymous
  function class inherits from js.ThisFunction.
* We add logic to handle repeated parameters in the treatment as
  well.

In addition, we remove the `@js.native` annotation from all the
"built-in" `js.FunctionN` and `js.ThisFunctionN` types. They lose
their prilileged status in the spec and in the compiler, to become
regular definitions that fit the above spec.