Handle inlining of Closures when the argument count does not match.
When there are too many arguments, we cancel, because we do not
have a good infrastructure to evaluate the side effects of the
extra arguments in the correct evaluation order. That case should
not be common anyway.

When there are not enough arguments, we fill the blanks with
`undefined` values, according to the JavaScript spec.