Reject fields with type NothingType in the IR checker.
A field with type `NothingType` cannot be valid IR, because there
is no valid zero value to initialize it with.

We introduce a deserialization hack that changes the types of such
fields to `NullType`, and correspondingly change its `Select`s to
throw when they are accessed.

We do not change the compiler in this commit yet, in order to test
backward compatibility.