Introduce a `Clone` IR node for the primitive object cloning op.
It replaces the user-land implementation in
`java.lang.ObjectClone`, which exploits undefined behavior.

Unlike `ObjectClone.clone(o)`, `<clone>(o)` is specified to work
for array values as well as any other `Cloneable`. This allows to
remove ugly hacks in `Infos` and `OptimizerCore` that treated
`clone()` specially when the receiver was/could be an array.

Other than supporting arrays, the implementation of `Clone(o)` is
identical to what was in `ObjectClone`, but in the `CoreJSLib`
instead.

We use a deserialization hack to replace the body of
`java.lang.Object.clone()` with the new implementation, which uses
the `Clone` node. This is similar to what we did for
`System.identityHashCode()` and the `IdentityHashCode` IR node.

In this commit, we do not change the original implementation, in
order to make sure that new linker works with an old standard
library.