Cache variables used 2+ times in RuntimeLong shift operations.
This is a practice we generally do in all the methods of
RuntimeLong, as explained in the big comment at the beginning of
the file. It was an oversight that we were not doing it in the
shift operations yet.

We never noticed it before because it rarely matters. This only
happens when we perform a shift by less than 32 directly on a Long
value that comes from outside the current method. In the entire
test suite, it seems that only the `z >>> 30` in
`SplittableRandom.mix64` is affected.