Refine the type of the receiver in an inlined method body.
When we inline a method, there are two pieces of information that
we gain about the receiver:

* It is non-null, and
* Its actual type must be at least as precise as the enclosing
  class of the target method.

The first information is "new" because NPEs are undefined behavior.
So when we see a call `x.m()`, even if we did not know that `x` is
non-null prior to the call, we know at that point that it is
because the alternative would be UB.

The enclosing class of the target method may be more specific than
the type of the receiver when we inline the only possible target of
a polymorphic dispatch.

---

In this commit, we refine the type of the `this` value within the
body of inlined methods using the above knowledge.

This has mostly two consequences:

When inlining a method that perform `this.isInstanceOf[Foo]`, we
can optimize it as `true` instead of `this$1 !== null`, if the
receiver is known to be a `Foo` at call site. This happens for
example when inlining `List.filter` (and a few more helpers that it
is itself delegates to).

The second consequence is a bit unexpected: when performing Long
operations, the receiver is now already declared as type
`RuntimeLong`, and this avoids an unnecessary `expandLong` in
`withNewLocalDef`. This reduces the amount of unnecessary temporary
variables that we emit for Long operations.

---

I was hoping that this would solve the `CharSequence` hack in
`FunctionEmitter`, but the current changes do not seem to be enough
for that.