Fix `checkClassType` for Scala 2.13.5.
`checkClassType` was renamed to `checkClassOrModuleType` upstream,
to be released in Scala 2.13.5, to account for its change in
behavior that was introduced in Scala 2.13.4. The renaming was done
in https://github.com/scala/scala/commit/53121483d349f61999f080272f330efd34cc786a

This commit introduces a compat hack to use the new method. This is
only in the hack for `classOf` desugaring, which corresponds to the
fact that `typedClassOf` upstream started using the renamed method
upstream.