Fix #4408: Compute the list of dispatchers that we need.
Instead of hard-coding the list of methods of hijacked classes that
need dispatchers, along with the hijacked classes that define them,
we compute the above information from the class hierarchy.

It turns out that the required information was already in the
global knowledge, except it was exposed in a weaker form. So this
change does not make incremental runs more sensitive.

These changes allow to get rid of the `CharSequence` hack in
`FunctionEmitter`, which was unsound.