Fix #4413: Handle JS symbols and bigints in `systemIdentityHashCode`.
By spec, for values that are not Scala objects nor primitives
corresponding to a hijacked class, `x.hashCode()` resolves to
`j.l.Object.hashCode()`, which delegates to
`System.identityHashCode()`.

Previously, the implementation of the latter would jump back to
`j.l.Object.hashCode()` for JS `bigint`s, and crash with a
`TypeError` for `symbol`s.

The main cause of these issues is that we thought we could delegate
back to `j.l.Object.hashCode()` for all primitives, because they
are supposed to be covered by hijacked classes. We had forgotten
`symbol`s, though, and anyway it does not work because not all JS
primitives are covered by a hijacked class.

In this commit, we break this cycle. `systemIdentityHashCode` never
delegates back to `j.l.Object.hashCode()` anymore. Instead, it
separately handles every type of primitive in JavaScript with a
dedicated hash code implementation. For strings and numbers, we
reuse `j.l.String.hashCode()` and `j.l.Double.hashCode()`, because
we can. For other values, we have an inline implementation.