Fix #4416: Remove stale methods from the Emitter's caches.
Previously, we forgot to reset the `_cacheUsed` flag of
non-static-like methods in the Emitter's caches. This caused them
never to be removed even if they were not used in a run.

The fix is beneficial anyway for memory consumption, but it should
not have been necessary for correctness. The issue happened because
the `IncOptimizer` can reuse past values of `version` for a given
method, if it was removed in a run and then reappeared. This
observation raises a larger question about whether that is valid
(and therefore should be accounted for in subsequent phases like
the Emitter), but that is left for a later fix.