Fix #4352: Unify error handling in java.util.Formatter.
Previously, we detected error conditions in `j.u.Formatter` in
separate code paths for each conversion. With that design, the
order in which we threw exceptions was not centrally managed, and
was not consistent with the behavior on the JVM.

In this commit, we significantly refactor the code paths in
`Formatter` to unify error handling for (almost) all conversions.
Only the `n` and `%` conversions have separate code paths; the
former out of necessity due to its spec, the latter out of
convenience and matching the JVM behavior.

For the unified path, we introduce an array
`ConversionsIllegalFlags` that maps each lowercase conversion
letter to its invalid flags. We naturally include `Uppercase` in
the bitset since we were already treating it as a flag elsewhere.
We also add `Precision`, which is not otherwise treated as a flag.

Using that array, we can uniformly detect virtually all error
conditions in a uniform way, before dispatching to the code of the
appropriate conversion.

There is one exception for the `o`, `x` and `X` conversions. They
have some flags that are only valid for `BigInteger` arguments.
Those require an additional check if we provide an `Int` or `Long`
argument.