Fix #4035: Correctly handle values close to midpoints in `parseFloat`.
Previously, we implemented `parseFloat(s)` as
`parseDouble(s).toFloat`. This is however not correct when the
exact value represented by `s` is close to a Float midpoint:
parsing it as a Double may round to the midpoint itself, which will
then be incorrectly rounded to the even float, even if the orginal
value was closer to the odd float.

We fix this in two different ways for the decimal and hexadecimal
notations.

For the hexadecimal notation, which is easier, we simply
artificially drop more precision when parsing as a Double, which
will force values very close but not equal to midpoints to be
rounded away from the midpoints.

For the decimal notation, this is much trickier. In fact, as proven
in the paper "How to Read Float Point Numbers Accurately" by
William D. Clinger, there is no solution that does not require big
integer arithmetic at some point. We take inspiration from the
`AlgorithmR` from that paper, which takes an initial value "close"
to the best approximation and improves it by 1 ULP. Since we
already have a close approximation (one that is at most 1 ULP away
from the best one), we can use that. However, we can dramatically
simplify the algorithm because we can leverage Double arithmetics
to parse only a Float. In particular, we can accurately compute and
represent the two adjacent Floats that enclose the best
approximation, as well as the midpoint between those, which is a
Double. The only real thing we keep from the paper is how to
accurately compare that midpoint with the exact value represented
by the string. This allows us to decide whether we need to round
up, down, or break a tie to even.