Fix #3706: Use a new instance for the result of divideToIntegralValue.
Previously, we reused the instance `integralValue` and reset its
fields. However, we had forgotten to reset `_precision` which
caused an internally inconsistent instance to leak to callers.

We could have fixed it just by resetting `_precision`, but this is
not something that is usually done elsewhere in the implementation
of `BigDecimal`. Instead, we simply return a brand new instance,
which will not be tainted by previous calculations.