Base stat/expr checking on the type system in IRChecker
Instead of distinguishing statements and expressions, we typecheck
everything in the same way. Statements are just regular trees whose
type happens to be NoType.

We use this opportunity to re-order the match statements in the same
order as the trees appear in Trees.scala.

This change uncovered issues in the way we emit If, Match and
TryCatch nodes in the compiler. Some of those nodes, when in
statement position, were typed as something else than NoType, but
had children typed as NoType. The previous IR checker did not flag
those cases as errors, whereas the new one does. We introduce a
deserialization hack to fix the type of If, Match and TryCatch
nodes in statement position to always be NoType.