Fix #3202: Format doubles exactly like on the JVM.
We address the differences by implementing the notations ourselves,
based on a decimal representation of numbers. We convert the Double
values to our Decimal representation based on the native
`toString()` method, which guarantees that its result contains as
few digits in the significand as necessary. Once we have a
`Decimal` value, we can round/rescale it as required, using the
round-half-up strategy, and finally format it.

We use the opportunity to reorder the notation functions to be in
the same order as the conversions they implement.