Bump the version to 1.6.0-SNAPSHOT and the IR version to 1.6-SNAPSHOT.
To accommodate the upcoming new configuration options in
`ESFeatures`.

We bump the IR version number as well because we will add a field
to the `$linkingInfo` structure produced by the linker. In order
for the newer version of the library to be able to rely on it, we
need the guarantee that it gets linked with a newer linker.