Allow to specify target versions of ECMAScript > 2015.
We introduce a new setting ESFeatures.esVersion to select specific
ES versions to target for compliance. Values of `ESVersion` are
defined from ES 5.1 up to ES 2020.

The setting `useECMAScript2015` is remapped to be the same as
`esVersion >= ESVersion.ES2015`. Since it represented two purposes
(relying on ES 2015 features, and enabling the ES 2015 semantics of
Scala.js language features), we introduce a separate setting
`useECMAScript2015Semantics`, and deprecate `useECMAScript2015` in
favor of either `esVersion` or `useECMAScript2015Semantics`. We
adapt and clean up the linker code to reflect those two purposes.

For now, we do not allow `useECMAScript2015Semantics` to be
directly configured. It is always derived from
`esVersion >= ESVersion.ES2015`. We do this not to introduce
unnecessary configuration for now, but we could make it
configurable on its own in the future.

We leverage `esVersion >= ESVersion.ES2017` in one place in the
core JS lib. More tests like that could be introduced in the
future.

We add `esVersion` in the generated `BuildInfo` for the test suite,
and use it to better constrain some our tests of JS features.

The `esVersion` is also made available to user-space libraries
through `LinkingInfo.esVersion`. `assumeES6` is deprecated in favor
of `esVersion`. We also introduce
`LinkingInfo.useECMAScript2015Semantics`, to match the setting from
`ESFeatures`.

We enhance `NodeJSEnvForcePolyfills` and our Jenkins scripts to
test with various target versions of ECMAScript.