Fix #4466: Use a dedicated function for Long-to-Float conversions.
The previous approach of first converting to Double then converting
the Double to Float was not correct for values close to Float
midpoints.

That encoding was baked in the IR, so we have to introduce a new
`UnaryOp.LongToFloat` operation, which we compile as calling
`RuntimeLong.toFloat`.

We then fix the implementation of `toFloat` to correctly handle
values that are close to Float midpoints, using a strategy similar
to what we had used in `parseFloat` for the hexadecimal format.