Regex: Store start and end for the indices of matched groups.
And `undefined` for unmatched groups. In addition, store the start
and end of the whole match at index 0.

The indices array is cached directly as a field `indices` on the
`MatchResult`. This allows the cache to be better shared between a
`Matcher` and a `MatchResult`.

This corresponds to the format of the `indices` field of the
`regexp-match-indices` ECMAScript proposal, scheduled for inclusion
in ECMAScript 2022.