Regex: Reuse the same RegExp for all the Matcher's of a Pattern.
Previously, each `Matcher` had its own copy of a "blueprint"
`RegExp`, to isolate its `lastIndex` state.

We now explicitly store that state as a separate `position` field
in `Matcher`, which we assign to `lastIndex` before `exec`, and
read from `lastIndex` after `exec`.

This allows to reuse the same `RegExp` for all matchers of a given
`Pattern`, and to confine its use to `Pattern.execFind()`.