Regex: Turn CharSequence into Strings upon entry in our implementation.
Since we have to convert the input `CharSequence`s to `String`s
anyway to be able to give them to `js.RegExp`, it is better to do
that upon entry into our implementation. We can force the
`toString()`s to be at the call site, which can then be inlined and
removed when the arguments are already Strings, which should be the
vast majority of use cases.