Upgrade to JUnit 4.13.2 and its `assertThrows` method.
This unfortunately creates zillions of clashes between
`org.junit.Assert.assertThrows` and our own
`AssertThrows.assertThrows`. The latter is more Scala friendly, so
we want to keep it. We use named imports when we use it to force it
over the one coming from `org.unit.Assert._`.

This allows to re-enable some tests in the scalaTestSuite in recent
versions of Scala, at the cost of having to disable them (and some
others) in older versions.

In addition, JUnit 4.13 brings some changes in how failed
assumptions are reported. This triggers changes in the JUnit Test
Outputs files as well as in our implementation.