JUnit: Add Int overloads to assertEquals and assertNotEquals.
JUnit has overloads of `assertEquals` and `assertNotEquals` that
take `Long`s and `Double`s as parameters, but not the other
primitive types. That means that when calling those methods with
`Int`s (or `Byte`s, `Short`s, `Char`s), the overload with `Long`s
is selected. While correct, this incurs a significant cost in
Scala.js.

We therefore introduce overloads taking `Int`s, although they are
not part of the JVM API.

The size of the fastopt.js file for our test suite is reduced by
40% (!) with this change. I could not measure a significant
difference in terms of link time or execution time, however.