Keep partest only for the latest minor versions of Scala.
In the past several years, I cannot recall a single time where a
bug was discovered in partest for a minor version of Scala and not
at the same time by the latest minor version. Therefore, it seems
those tests are pure load (human maintenance and machine time)
without any benefit.

This commit removes partest for Scala 2.12.{1-12}, keeping only
2.11.12, 2.12.13 and 2.13.5.