Fix #4496: Add js.import.meta to access ES2020's `import.meta`.
Since in ECMAScript, `import.meta` is a dedicated syntactic form,
we add a new IR node for this meta-property. We make it accessible
at the language level through

  js.`import`.meta

It is a linking error to reference `import.meta` when the module
kind is not `ESModule`, since attempting to use it in a Script
results in an early SyntaxError.