Enable JVM codegen and disable ExternalCompile.
Now that we got rid of the idiosynchrasies of our build that
prevented us from using the JVM back-end, and hence sbt's
integration with the compiler, we can enable them.

We configure the incremental compiler to recompile everything or
nothing on the javalib and scalalib projects, because the
incremental compiler can still get confused about those.