Upgrade our build to sbt 1.5.4.
As a side effect, this solves the issue that testSuite2_*/test was
always logging debug messages. Now they are only displayed with the
`-v` option. This will dramatically reduce the disk space used by
logs on the CI servers.

The sbt plugin is still compiled against sbt 1.0.0 artifacts, and
the scripted tests are still executed against sbt 1.2.8. Hopefully
that is enough to ensure that we do not break sbt 1.{2,3,4}.x
users.