Avoid using the (?m) flag in the regexes in linker-interface.
The (?m) flag required ES2018. Avoiding it allows the
linker-interface (and linker) to work with earlier versions of
ECMAScript.

The regex `fileFieldRe` used the (?m) but for no purpose. We just
remove it.

The regex `sourceMapRe` did use the properties of (?m) for ^ and $.
We use `(?:^|\n)` and `(?:\n|$)` instead, and adjust the code using
the regex to account for the potential extra `\n`'s that are
matched.

While we are there, we fix potential performance issues in those
two regexes. We replace `.*x` by `[^x]*x` to avoid excessive
backtracking.