Fix #3563: Add config for a header comment to add to .js files.
This commits adds a new setting in `StandardConfig` to specify a
header comment to prepend to the generated .js files. It can be set
in sbt with

    scalaJSLinkerConfig ~= { _.withJSHeader("// foo\n") }

The header must satisfy the following properties:

* It must contain only valid JavaScript comments and whitespace.
* It cannot contain Unicode new line characters except NL ('\n').
* If non-empty, it must end with a new line.
* It must not contain unpaired surrogate characters.

This ensures that the generated code is still valid, and that the
`jsHeader` option cannot be abused to insert JavaScript code that
breaks the linker's abstractions.

As discussed on the issue, if someone *really* wants to insert a
shebang line in their generated .js file, it is possible to insert
an empty line with `withJSHeader("\n")`, and then post-process the
.js file to replace that empty line with a shebang line. That
would preserve the line numbers in the source maps.