Fix #3953: Patch the types of ParamDefs to match their method signature.
In some rare cases that involve Higher Kinded Types and type
aliases, scalac produces DefDef's whose params' types do not match
the method type. This is filed upstream as
https://github.com/scala/bug/issues/11884

We work around the issue by patching a posteriori the types of
`js.ParamDef`s and their `js.VarRef`s to match the type advertised
by the method type.

In the entire test suite, the only method that requires a patch is
the `PartialFunction`'s `applyOrElse` in the newly added test case.