Introducing `WeakNoModule`
The idea behind `WeakNoModule` is quite simple: let force `Analyzer` to
ignore checking of used modules from `.sjsir` which is linking.

It might be required when someone is tried to use inside web-project a
library which is universal and has checking to loading used modules like
`ws` in runtime.

Without `WeakNoModule` this project can't be linked because it fails at
linking stage with error like:
```
[error] BlaBla needs to be imported from module 'xxx' but module support is disabled
```