Fix #4542: OOM on many dynamicImports
We expose ir.SHA1 publicly so we can use it in the linker. Because we
allow binary breakages in ir in minor versions, this is not a big
concern.

Further, it turns out that this actually subtly improves on the
grouping behavior:

Take the following dependency graph:

```dot
digraph {
  # static edges
  main -> A1
  main -> A2
  A1 -> O1
  B1 -> O1
  A2 -> O2
  B2 -> O2

  # dynamic edges
  edge[style=dashed]
  A1 -> B1
  A2 -> B2
}
```

In the previous algorithm the resulting tags would have been:

- O1: `({main}, {B1})`
- O2: `({main}, {B2})`

As a result, `O1` and `O2` would get classified into different
internal modules unnecessarily.

In the new version the resulting "tags" are:

- O1: `({main}, {})`
- O2: `({main}, {})`

Therefore, `O1` and `O2` would get classified into the same internal
module.