Do not rely on RuntimeLong extending the same things that j.l.Long.
We will not be able to keep that promise for `Constable` and
`ConstantDesc`, since they won't be available when building
`RuntimeLong` on JDK < 12.

There was only one place in the linker that required this to be
true, and it was different from a bigint-based implementations
than for a RuntimeLong-based implem. Removing this requirement
actually simplifies the code of the linker.