Fix #4554: Correctly identify default accessors of native JS def params.
Previously, we tested whether the default accessor itself had the
`@js.native` annotation. That is however never the case. Instead,
we now test whether the associated method has the annotation.

We still emit the *definitions* of those default accessors, even
though that they are never called anymore, because of backward
binary compatibility. If we stopped emitting them, a library
compiled with a newer Scala.js might not link anymore against an
application built with a previous Scala.js.