Fix #4129: Warn if the default global execution context is used.
The behavior in the presence of imports is unfortunately not ideal
here, especially when the implicit ExecutionContext is
imported.

Ideally, we'd like a warning on the import already and the ability to
silence the warning once for an entire file. However, imports get
fully resolved in the typer phase, so this would require processing
imports in the pre-typer phase, where the entire suppression
infrastructure is not available (it requires typing of the @nowarn
annotations).

Therefore, we live with the sub-optimal situation, especially given
that there is a compiler flag to disable the warnings overall.