Fix #4560: Explicitly redirect JS envs outputs to System.{out,err}.
When running in server mode, sbt installs alternative streams for
`System.out` and `System.err`. However, those alternatives are not
guaranteed to be used when launching JSEnvs with `inheritOut` and
`inheritErr` set to `true`. In particular, a JSEnv that spawns
separate processes with the inheritIO mode will bypass
redirections set up with `System.set{Out,Err}`.

We now explicitly redirect the outputs of JS envs to `System.out`
and `System.err`, so that they always take the alternative streams
set up by sbt into account.