Fix most of #2638: Access `new.target` in JS class constructors.
This commit introduces a new primitive method `js.new.target`,
which is the direct equivalent of JavaScript's `new.target`
meta-property when used in a class constructor.

It is only legal when used directly in the constructor of a
non-native JS class (or object), including:

* in a statement,
* in the right-hand-side of a `val` or `var`,
* in a parameter to the super constructor,
* or in a secondary constructor.

It is never legal anywhere else, in particular:

* outside the constructor of a non-native JS class,
* in the right-hand-side of a `lazy val` or a `def`, nor
* in a lambda or by-name parameter.

In JavaScript, `new.target` can also be meaningfully used in a
non-arrow `function`. That use case is not addressed by this
commit at the Scala.js language level. The IR is however already
capable of dealing with the `JSNewTarget()` node in the body of
a non-arrow `Closure`.

When trying to link a `new.target` when targeting ES 5.1 results
in a linking error, as there is no reliable translation of
`new.target` to ES 5.1.