Fix #4596: Allow hashbang comments at the beginning of the jsHeader.
There is a Stage 3 proposal for ECMAScript to officially allow
hashbang comments at the very beginning of source files (scripts
and modules):
https://github.com/tc39/proposal-hashbang

According to the proposal, they are to be true comments from the
point of view of ECMAScript, so it makes sense to accept them in
`jsHeader`. The current spec text specifies that, after the
starting `#!`, they can contain the same characters as single-line
comments starting with `//`.