Fix #4601: Patch the type of Labeled nodes in the 1.5 deser hack.
The deserialization hack introduced in
3601272fd49d56d452e2b252d0ee171922031e00 sometimes created invalid
IR from valid IR. This happened when a Labeled block in statement
position was typed with something else than `NoType`, and contained
an `If` or `Match` node of that same type. The deserialization hack
would change the type of the `If` or `Match` to `NoType`, but not
the type of the enclosing `Labeled` node. This pattern of IR is
produced by Scala 3.0.x, which emits IR version 1.5.

We fix this issue by adding `Labeled` to the list of nodes whose
type is patched by the hack. As written in the elaborated comment,
this is the last of the 4 nodes that are control structures whose
type is explicitly specified. Only those need to be adapted,
because:

1. The other control structures directly get their type from their
   children, and are therefore valid by construction.
2. Non-control structures do not transfer their "status" of being
   in expression position or in statement position.

This is unfortunately not tested.We would have to either load IR
generated by Scala 3.0.0, or write hand-coded IR v1.5 and load it
back in the linker tests. We do not have the required
infrastructure for either of these solutions.