Fix #4610: Warn when `-scalajs` or `-Xplugin:scalajs-compiler.jar` is missing.
Often, this happens because a user accidentally used
`scalacOptions := ...` instead of using `++=`. The symptoms are
however obscure, and it can be difficult to diagnose.

We now explicitly detect this, and emit a warning if something is
amiss.

This feature is not automatically tested because `scripted` does
not have any feature to test that a specific log message is
written.